requirements.txt
pysocks
httpx
requests
colorama
aiohttp
colorama
requests
httpx
screen
progressbar
tqdm
multiprocess